// src/types/@gearbox-protocol/router/contracts/interfaces/IClosePathResolver.ts
export type BalanceStruct = { token: string; balance: BigNumberish };