// src/types/@gearbox-protocol/router/contracts/interfaces/IRouter.ts
export type SwapTaskStruct = {
  swapOperation: BigNumberish;
  creditAccount: string;
  tokenIn: string;
  tokenOut: string;
  connectors: string[];
  amount: BigNumberish;
  slippage: BigNumberish;
  externalSlippage: boolean;
};

export type SwapTaskStructOutput = [
  number,
  string,
  string,
  string,
  string[],
  BigNumber,
  BigNumber,
  boolean
] & {
  swapOperation: number;
  creditAccount: string;
  tokenIn: string;
  tokenOut: string;
  connectors: string[];
  amount: BigNumber;
  slippage: BigNumber;
  externalSlippage: boolean;
};

export type MultiCallStruct = { target: string; callData: BytesLike };

export type MultiCallStructOutput = [string, string] & {
  target: string;
  callData: string;
};

export type RouterResultStruct = {
  amount: BigNumberish;
  gasUsage: BigNumberish;
  calls: MultiCallStruct[];
};

export type RouterResultStructOutput = [
  BigNumber,
  BigNumber,
  MultiCallStructOutput[]
] & { amount: BigNumber; gasUsage: BigNumber; calls: MultiCallStructOutput[] };

export type PathOptionStruct = {
  target: string;
  option: BigNumberish;
  totalOptions: BigNumberish;
};

export type PathOptionStructOutput = [string, number, number] & {
  target: string;
  option: number;
  totalOptions: number;
};

export type BalanceStruct = { token: string; balance: BigNumberish };

export type BalanceStructOutput = [string, BigNumber] & {
  token: string;
  balance: BigNumber;
};
// ...
export interface IRouter extends BaseContract {
// ...
  callStatic: {
    componentAddressById(
      pf: BigNumberish,
      overrides?: CallOverrides
    ): Promise<string>;

    findAllSwaps(
      swapTask: SwapTaskStruct,
      overrides?: CallOverrides
    ): Promise<RouterResultStructOutput[]>;

    findBestClosePath(
      creditAccount: string,
      connectors: string[],
      slippage: BigNumberish,
      pathOptions: PathOptionStruct[],
      iterations: BigNumberish,
      force: boolean,
      overrides?: CallOverrides
    ): Promise<
      [RouterResultStructOutput, BigNumber] & {
        result: RouterResultStructOutput;
        gasPriceTargetRAY: BigNumber;
      }
    >;

    findOneTokenPath(
      tokenIn: string,
      amount: BigNumberish,
      tokenOut: string,
      creditAccount: string,
      connectors: string[],
      slippage: BigNumberish,
      overrides?: CallOverrides
    ): Promise<RouterResultStructOutput>;

    findOpenStrategyPath(
      creditManager: string,
      balances: BalanceStruct[],
      target: string,
      connectors: string[],
      slippage: BigNumberish,
      overrides?: CallOverrides
    ): Promise<[BalanceStructOutput[], RouterResultStructOutput]>;

    getGasPriceTokenOutRAY(
      token: string,
      overrides?: CallOverrides
    ): Promise<BigNumber>;

    isRouterConfigurator(
      account: string,
      overrides?: CallOverrides
    ): Promise<boolean>;

    tokenTypes(arg0: string, overrides?: CallOverrides): Promise<number>;

    version(overrides?: CallOverrides): Promise<BigNumber>;
  };
// ...
}